moved sidemenu to its own folder
